separated tests classes into two files
